implement simple lookup REST API
We will dedicate the route /api for all REST API calls, which produce
JSON. This commit implements /api/lookup/<word>. If the lookup fails (no
word matches the lookup string), then an empty object is returned.
Otherwise, a JSON dump of the identified Entry (as defined in models.py)
is returned.